requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
django-qrcode==0.3
Flask==3.1.1
ImageHash==4.3.2
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pycparser==2.22
PyWavelets==1.8.0
qrcode==8.2
scikit-image==0.25.2
scipy==1.15.3
tifffile==2025.5.10
Werkzeug==3.1.3
